Split graphql modules into two packages
