+++ plugins/graphql-catalog/.eslintrc.js
