+++ plugins/graphql-catalog/codegen.yml
@@ -1,0 +1,16 @@
+overwrite: true
+schema: "src/schema.ts"
+generates:
+  ./src/modules:
+    preset: graphql-modules
+    presetConfig:
+      baseTypesPath: ../__generated__/graphql.ts
+      filename: __generated__/types.ts
+    plugins:
+      - add:
+          content: '/* eslint-disable */'
+      - typescript
+      - typescript-resolvers
+  ./__generated__/schema.graphql:
+    plugins:
+      - schema-ast
