+++ plugins/graphql-catalog/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@backstage/plugin-graphql-backend",
-  "description": "An experimental Backstage backend plugin for GraphQL",
-  "version": "0.1.33-next.1",
+  "name": "@backstage/plugin-graphql-catalog",
+  "description": "An experimental Backstage GraphQL catalog module",
+  "version": "0.3.17-next.1",
   "main": "src/index.ts",
   "types": "src/index.ts",
   "license": "Apache-2.0",
@@ -10,6 +10,10 @@
     "main": "dist/index.cjs.js",
     "types": "dist/index.d.ts"
   },
+  "exports": {
+    "./schema.graphql": "./src/catalog/catalog.graphql",
+    "./package.json": "./package.json"
+  },
   "backstage": {
     "role": "backend-plugin"
   },
@@ -17,7 +21,7 @@
   "repository": {
     "type": "git",
     "url": "https://github.com/backstage/backstage",
-    "directory": "plugins/graphql-backend"
+    "directory": "plugins/catalog-graphql"
   },
   "keywords": [
     "backstage",
@@ -30,33 +34,29 @@
     "test": "backstage-cli package test",
     "prepack": "backstage-cli package prepack",
     "postpack": "backstage-cli package postpack",
+    "generate:types": "graphql-codegen",
     "clean": "backstage-cli package clean"
   },
   "dependencies": {
     "@backstage/backend-common": "workspace:^",
-    "@backstage/catalog-client": "workspace:^",
-    "@backstage/config": "workspace:^",
     "@backstage/plugin-graphql-common": "workspace:^",
-    "@types/express": "^4.17.6",
-    "apollo-server": "^3.0.0",
-    "apollo-server-express": "^3.0.0",
-    "express": "^4.17.1",
-    "express-promise-router": "^4.1.0",
-    "graphql": "^16.0.0",
+    "@graphql-tools/load-files": "^6.5.3",
+    "@graphql-tools/utils": "^8.6.5",
     "graphql-modules": "^2.0.0",
-    "helmet": "^6.0.0",
-    "reflect-metadata": "^0.1.13",
-    "winston": "^3.2.1",
-    "yn": "^4.0.0"
+    "graphql-type-json": "^0.3.2"
   },
   "devDependencies": {
     "@backstage/cli": "workspace:^",
-    "@types/supertest": "^2.0.8",
-    "msw": "^1.0.0",
-    "supertest": "^6.1.3"
+    "@graphql-codegen/add": "^4.0.0",
+    "@graphql-codegen/cli": "^3.0.0",
+    "@graphql-codegen/graphql-modules-preset": "^3.0.0",
+    "@graphql-codegen/schema-ast": "^3.0.0",
+    "@graphql-codegen/typescript": "^3.0.0",
+    "@graphql-codegen/typescript-resolvers": "^3.0.0"
   },
   "files": [
     "dist",
-    "schema.gql"
+    "src/**/*.graphql",
+    "README.md"
   ]
 }
