+++ plugins/graphql-catalog/src/catalog/catalog.graphql
