+++ plugins/graphql-common/.gitignore
@@ -1,0 +1,1 @@
+__generated__
