+++ plugins/graphql-common/CHANGELOG.md
