+++ plugins/graphql-common/package.json
@@ -1,6 +1,6 @@
 {
-  "name": "@backstage/plugin-catalog-graphql",
-  "description": "An experimental Backstage catalog GraphQL module",
+  "name": "@backstage/plugin-graphql-common",
+  "description": "An experimental Backstage GraphQL common module",
   "version": "0.3.19-next.1",
   "main": "src/index.ts",
   "types": "src/index.ts",
@@ -52,7 +52,6 @@
     "graphql": "^16.0.0",
     "graphql-modules": "^2.0.0",
     "graphql-relay": "^0.10.0",
-    "graphql-type-json": "^0.3.2",
     "lodash": "^4.17.21"
   },
   "devDependencies": {
@@ -60,8 +59,6 @@
     "@backstage/test-utils": "workspace:^",
     "@effection/jest": "^2.0.4",
     "@faker-js/faker": "^7.2.0",
-    "@frontside/graphgen": "^1.8.1",
-    "@frontside/graphgen-backstage": "0.3.0",
     "@graphql-codegen/add": "^4.0.0",
     "@graphql-codegen/cli": "^3.0.0",
     "@graphql-codegen/graphql-modules-preset": "^3.0.0",
@@ -72,7 +69,7 @@
   },
   "files": [
     "dist",
-    "src/app/modules/**/*.graphql",
+    "src/**/*.graphql",
     "README.md"
   ]
 }
