+++ plugins/graphql-common/src/core/core.graphql
