+++ plugins/graphql-common/src/schema.ts
@@ -13,6 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import { transformSchema } from './service/transform';
+import { printSchemaWithDirectives } from '@graphql-tools/utils';
+import { transformSchema } from './transformSchema';
 
-export default transformSchema([]);
+export default printSchemaWithDirectives(transformSchema([]));
