+++ plugins/graphql-common/src/types.ts
